feat(refactor): extract URL synchronization logic into headless stores (#602)
