+++ src/routes/calculator/+page.svelte
@@ -49,10 +49,10 @@
     loadAdvancedOptionsFromStorage,
     persistAdvancedOptions,
   } from "$lib/state/advanced-options.svelte";
-  import {
     createInverseLookupState,
     type InverseLookupState,
   } from "$lib/state/inverse-lookups.svelte";
+  import { setupCalculatorUrlSync } from "$lib/state/calculator-url-sync.svelte.ts";
   import { HIGH_E_SIDE, LOW_E_SIDE } from "$lib/utils/inverse-stp";
   import { LibdedxError, type InverseCsdaResult, type EnergyUnit } from "$lib/wasm/types";
   import { negotiateVersion } from "$lib/utils/url-version.js";
@@ -447,111 +447,15 @@
     calcState.triggerCalculation();
   });
 
-  $effect(() => {
-    // Read advOptsKey to establish reactive dependency on nested changes
-    const _advOptsKey = advOptsKey;
-    void _advOptsKey;
-
-    if (!urlInitialized || !calcState || !entityState) return;
-    // In advanced mode, wait for multiProgState to be initialized so the URL update
-    // does not overwrite the reloaded URL (which may contain programs= from a previous
-    // session) before the multiProgState effect has had a chance to read it.
-    if (isAdvancedMode.value && multiProgState === null) return;
-
-    // Build inverse mode state for URL encoding
-    let inverseModeState: InverseModeUrlState | undefined;
-    if (inverseLookupState && isAdvancedMode.value) {
-      if (inverseLookupState.activeTab === "csda") {
-        inverseModeState = {
-          imode: "csda",
-          lookups: inverseLookupState.rangeRows
-            .filter((r) => r.text.trim() !== "")
-            .map((r) => {
-              const trimmed = r.text.trim();
-              const numeric = trimmed.match(/^([\d.eE+-]+)/)?.[1] ?? trimmed;
-              return {
-                rawInput: r.unitFromSuffix ? numeric : trimmed,
-                unit: r.unit,
-                unitFromSuffix: r.unitFromSuffix,
-              };
-            }),
-          iunit: inverseLookupState.rangeMasterUnit,
-        };
-      } else if (inverseLookupState.activeTab === "stp") {
-        inverseModeState = {
-          imode: "stp",
-          lookups: inverseLookupState.stpRows
-            .filter((r) => r.text.trim() !== "")
-            .map((r) => ({
-              rawInput: r.text.trim(),
-              unit: r.unit,
-              unitFromSuffix: false,
-            })),
-          iunit: inverseLookupState.stpMasterUnit,
-        };
-      }
-    }
-
-    const selectedMaterial = entityState.selectedMaterial;
-    // Narrow to built-in MaterialEntity; ExternalOnlyMaterial lacks isGasByDefault
-    const builtinMaterial =
-      selectedMaterial && "isGasByDefault" in selectedMaterial ? selectedMaterial : null;
-    const customUrlFields = isCustomMaterial(builtinMaterial)
-      ? customMaterialUrlFields(builtinMaterial)
-      : {};
-
-    const selectedParticleId = entityState.selectedParticle?.id;
-    const activeMultiProgramState = isAdvancedMode.value ? multiProgState : null;
-
-    const urlState = {
-      particleId: typeof selectedParticleId === "number" ? selectedParticleId : null,
-      materialId:
-        builtinMaterial && typeof builtinMaterial.id === "number" ? builtinMaterial.id : null,
-      // External program IDs are not yet URL-encoded in programId; null means auto-select
-      programId:
-        typeof entityState.resolvedProgramId === "number" ? entityState.resolvedProgramId : null,
-      rows: calcState.rows,
-      masterUnit: calcState.masterUnit,
-      externalSources: loadedExternalSources,
-      ...customUrlFields,
-      // Include advanced mode state when active
-      ...(activeMultiProgramState
-        ? {
-            isAdvancedMode: true,
-            // Emit ALL selected programs in display order (default program first)
-            // so the URL is the canonical full list and consumers can reconstruct
-            // the complete comparison without needing to infer the default.
-            selectedProgramIds: activeMultiProgramState.selectedProgramIds,
-            quantityFocus: activeMultiProgramState.quantityFocus,
-            // Include advanced options when in advanced mode
-            advancedOptions: advancedOptions.value,
-            materialIsGas: builtinMaterial?.isGasByDefault,
-          }
-        : {}),
-      // Include inverse mode state when active
-      ...(inverseModeState || {}),
-      // Encode istpbranch when STP column visibility is "both"
-      ...(inverseModeState?.imode === "stp" && inverseLookupState?.stpBranchState === "both"
-        ? { istpBranchState: "both" as const }
-        : {}),
-    };
-    // Use calculatorUrlQueryString so `:` and `,` are written literally
-    // (RFC 3986 §3.4 permits them unencoded in the query component).
-    // This matches the format the browser stores in window.location.search
-    // and keeps URLs human-readable (e.g. `energies=100,500:keV`).
-    //
-    // Build from `window.location.pathname` rather than `page.url.pathname`
-    // so reading `page.url` does not register a reactive dependency on the
-    // very URL we are about to rewrite — otherwise this effect re-runs on
-    // every `replaceState` and forms a (silent) replaceState loop.
-    const queryString = calculatorUrlQueryString(urlState);
-    const next = `${window.location.pathname}?${queryString}`;
-    if (next === `${window.location.pathname}${window.location.search}`) return;
-    // Use untrack so reading page.state does not register a reactive dependency.
-    // Without this, replaceState updates the SvelteKit page store (new object
-    // reference for page.state) which re-triggers this effect on every call.
-    untrack(() => replaceState(next, page.state));
-  });
+  setupCalculatorUrlSync(
+    () => calcState,
+    () => entityState,
+    () => inverseLookupState,
+    () => multiProgState,
+    () => urlInitialized,
+    () => loadedExternalSources,
+    () => advOptsKey,
+  );
 
   $effect(() => {
     if (calcState && entityState?.isComplete) {
